Projekat 2/ROS2/requirements.txt
pyserial
ahrs
numpy
websockets
